backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
google-genai==1.31.0
python-dotenv==1.1.1

